requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.145
botocore==1.34.145
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.14
django-admin-rangefilter==0.12.0
django-import-export==3.3.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-jazzmin==3.0.0
django-postgres-extra==2.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docstring_parser==0.16
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
jmespath==1.0.1
loguru==0.7.2
multidict==6.0.5
openai==1.40.3
packaging==24.1
pdfkit==1.0.0
pinecone-client==2.2.2
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
tiktoken==0.7.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain_google_vertexai==2.0.1
langfuse==2.50.2
langgraph==0.2.23
llama-index==0.11.12
llama-index-graph-stores-neo4j==0.3.2
openmeter==1.0.0b184
wikipedia==1.4.0
python-json-logger==2.0.7
wrapt==1.16.0
yarl==1.9.7
zipp==3.20.1
sentry-sdk==2.14.0
django-redis==5.4.0
whyhow==0.1.5
adrf==0.1.8
twilio==9.3.6
websockets==12.0
SpeechRecognition==3.11.0
daphne==4.1.2
channels==4.1.0
google-cloud-speech==2.28.1
pydub==0.25.1
mysql-connector==2.2.9
kafka-python==2.0.2
ulid-py==1.1.0
mysql-connector-python==9.1.0
celery==5.4.0
django-defender==0.9.8
pinecone[grpc]==5.4.2
django-health-check==3.18.3
kafka-healthcheck==0.0.1
google-cloud-bigquery==3.25.0
google-cloud-texttospeech==2.23.0
langchain-google-genai==2.0.6
django-google-sso==8.0.0
qdrant-client==1.12.2
psutil==6.1.1
firebase-admin==6.6.0
newrelic==10.8.1
